Extract more Blog files to feature folder
