+++ src/Features/Blog/Containers/newsContainer/index.js
@@ -1,0 +1,9 @@
+import React from 'react'
+// import PropTypes from 'prop-types'
+// import styled from 'styled-components'
+
+const NewsContainer = () => {
+  return <div>NEWS</div>
+}
+
+export default NewsContainer
