--- src/containers/newsContainer/index.js
@@ -1,13 +1,0 @@
-import React, { Fragment } from 'react'
-import PropTypes from 'prop-types'
-import styled from 'styled-components'
-
-const NewsContainer = () => {
-  return (
-    <div>
-    NEWS
-    </div>
-  )
-}
-
-export default NewsContainer
