+++ src/pages/blog.js
@@ -1,11 +1,11 @@
 import React, { Fragment } from 'react'
-import styled from 'styled-components'
-import { Container, Row, Column } from '../components/grid'
-import { media } from '../theme/media'
+// import styled from 'styled-components'
+// import { Container, Row, Column } from '../components/grid'
+// import { media } from '../theme/media'
 import ImageBanner from '../components/imageBanner'
 import theme from '../theme/theme'
-import ViewsContainer from '../containers/viewsContainer'
-import NewsContainer from '../containers/newsContainer'
+import ViewsContainer from '../Features/Blog/Containers/viewsContainer'
+import NewsContainer from '../Features/Blog/Containers/newsContainer'
 
 const Blog = () => (
   <Fragment>
